+++ lib/dexter.rb
@@ -12,6 +12,7 @@ require "dexter/indexer"
 require "dexter/log_parser"
 require "dexter/csv_log_parser"
 require "dexter/pg_stat_activity_parser"
+require "dexter/sql_parser"
 require "dexter/processor"
 require "dexter/query"