+++ lib/dexter/processor.rb
@@ -13,6 +13,8 @@ module Dexter
           PgStatActivityParser.new(@indexer, @collector)
         elsif options[:input_format] == "csv"
           CsvLogParser.new(logfile, @collector)
+        elsif options[:input_format] == "sql"
+          SqlParser.new(logfile, @collector)
         else
           LogParser.new(logfile, @collector)
         end